fix(build): four warnings, one of which was load-bearing
A clean Debug build with `-Wall -Wextra -Wshadow -Wdouble-promotion` had four
warnings in this project's own code. Three were cosmetic and one was not.

`RoadBuilder::buildMarkings` binds `side` to the side street's *material* and
then, forty lines later, opens `for (const float side : {-1.0f, 1.0f})` inside
it. Both lines of the loop body read the float, which is what they wanted, so
it compiled and it was right — and it was right by accident. Renaming the loop
variable to `lane` turned the second line into a compile error against a
`Material*`, which is precisely the failure the shadow was hiding and precisely
what the next person to move either declaration would have hit. Two things
called side in one function, one of them a sign and one of them a surface.

The other three: an unused loop index where a count was wanted, and two implicit
float-to-double promotions in the frame profile and the signage sizing.

Also here: the content set rebuilt from scratch (202 surfaces, 11 models), the
screenshot sets re-captured from the final tree, and the flagship five
re-rendered at 1920 x 1080. `validate-assets` is clean, both build types are
clean, and all nine test suites pass in Debug and Release.

+++ street/src/StreetApplication.cpp
@@ -488,12 +488,12 @@ void StreetApplication::recordFrame()
     if (frameBudget_ <= 0 || framesDrawn_ <= kProfileWarmup) return;
     const SceneRenderer::Stats& stats = renderer_->stats();
     profile_.frameMs.push_back(stats.frameMs);
-    profile_.cullMs    += stats.cullMs;
-    profile_.shadowMs  += stats.shadowMs;
-    profile_.prepassMs += stats.prepassMs;
-    profile_.skyMs     += stats.skyMs;
-    profile_.opaqueMs  += stats.opaqueMs;
-    profile_.postMs    += stats.postMs;
+    profile_.cullMs    += static_cast<double>(stats.cullMs);
+    profile_.shadowMs  += static_cast<double>(stats.shadowMs);
+    profile_.prepassMs += static_cast<double>(stats.prepassMs);
+    profile_.skyMs     += static_cast<double>(stats.skyMs);
+    profile_.opaqueMs  += static_cast<double>(stats.opaqueMs);
+    profile_.postMs    += static_cast<double>(stats.postMs);
     profile_.draws       += stats.drawCalls;
     profile_.shadowDraws += stats.shadowDrawCalls;
     profile_.triangles   += static_cast<long long>(stats.triangles);
